+++ lib/spark/component.rb
@@ -1,21 +1,21 @@
 # frozen_string_literal: true
 
-require_relative "element"
+require_relative "component/element"
 
 if defined?(ActionView::Component::Base)
-  require "spark/integration/action_view_component"
+  require_relative "component/integration/action_view_component"
 end
 
 module Spark
   module Component
     def self.included(base)
-      base.include Spark::Element unless base < Spark::Element
+      base.include Spark::Component::Element unless base < Spark::Component::Element
 
       # If an Integration is defeind include its modules if the component extends
       # the defined base class
-      return unless defined?(Spark::Integration)
+      return unless defined?(Spark::Component::Integration)
 
-      base.include(Spark::Integration::Component) if base < Spark::Integration.base_class
+      base.include(Spark::Component::Integration) if base < Spark::Component::Integration.base_class
     end
   end
 end
