+++ test/spark/attr_hash_test.rb
@@ -1,34 +1,34 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "spark/tag/hash"
+require "spark/component/attr_hash"
 
 module Spark
-  module Tag
-    class HashTest < Minitest::Test
+  module Component
+    class AttrHashTest < Minitest::Test
       parallelize_me!
 
       def test_html_attrbute_format_on_to_s
-        hash = Tag::Hash.new
+        hash = AttrHash.new
         hash.add foo: :bar
 
         assert_equal %(foo="bar"), hash.to_s
       end
 
       def test_prefix_prepends_keys_on_to_s
-        data = Tag::Hash.new(prefix: :data)
+        data = AttrHash.new(prefix: :data)
         data.add foo_bar: :baz
 
         assert_equal %(data-foo-bar="baz"), data.to_s
 
-        aria = Tag::Hash.new(prefix: :aria)
+        aria = AttrHash.new(prefix: :aria)
         aria.add foo_bar: :baz
 
         assert_equal %(aria-foo-bar="baz"), aria.to_s
       end
 
       def test_dasherize_keys
-        hash = Tag::Hash.new
+        hash = AttrHash.new
 
         hash.add(foo_bar: :baz)
