+++ test/spark/class_attr_test.rb
@@ -1,22 +1,22 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "spark/tag/classname"
+require "spark/component/attr_class"
 
 module Spark
-  module Tag
-    class ClassnameTest < Minitest::Test
+  module Component
+    class ClassAttrTest < Minitest::Test
       parallelize_me!
 
       def test_classname_to_s
-        classes = Classname.new
+        classes = AttrClass.new
         classes.add :foo, :bar, :baz
 
         assert_equal("foo bar baz", classes.to_s)
       end
 
       def test_classnames_are_unique
-        classes = Classname.new
+        classes = AttrClass.new
         classes.add :foo, :bar, :baz, :bar
         classes.add :foo
         assert_equal "foo bar baz", classes.to_s
@@ -26,7 +26,7 @@ module Spark
       end
 
       def test_classnames_track_base_class_separately
-        classes = Classname.new
+        classes = AttrClass.new
 
         classes.add :foo, :bar
         assert_nil classes.base
